ai-agent/requirements.txt
langgraph
langgraph[cli]
langchain
langchain-google-genai
pyPDF2
python-multipart
python-dotenv
typing-extensions
pydantic
fastapi
uvicorn
